requirements.txt
sqlalchemy
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
python-dotenv==1.0.0
Werkzeug==3.0.6
requests==2.31.0
cryptography==44.0.1
itsdangerous==2.1.2